Allow building GDB natively on Morello
This adds a NATIVE_NON_PURECAP target info (which is only used for GDB)
to pass -mabi=aapcs to the compiler. We also override PKG_CONFIG_LIBDIR
to find the AArch64 gmp/expat libraries in /usr/local64 since otherwise
the build will fail when trying to link against purecap libraries.