Simplify handling of autogen.sh
If configure does not exist but autogen.sh does, try running that first
as part of the configure process. Some projects (the X11 libraries)
./autogen.sh will run ./configure automatically afterwards, but we
don't want to do that since we are using a separate build dir and not
passing the necessary configure arguments to autogen.sh. Fortunately,
this behaviour can be changed by setting NOCONFIGURE in the environment.