Default to using "make check" as the test step for autotools
Many autotools projects allow you to run the tests using "make check".
We now use "make check -n" to see if the target exists.W with GNU make
we could actually check the exit code of "make -q check", but
this is not portable to bmake, so just use "-n".