Fix incorrect target being built for shared-mime-info
When building kcoreaddons-aarch64, the call to
get_update_mime_database_path() uses a SharedMimeInfo (aarch64) instance,
so cls.target is shared-mime-info-aarch64 which means it will will not use
the cheribuild_xtarget since the target name already includes the
architecture. Fix this by using native_instance.target and adding an
assertion. This should fix the errors seen in Jenkins.