Don't allow dependency_error() to execute other targets before setup()
We don't want to trigger runs of other targets before setup()
completes since this could cause those projects to access a partially
initialized class if there are cyclic dependencies.

With a bit more refactoring we can turn this into a fatal error, or
possibly defer the execution (although that could be awkward).