run_cheribsd_tests: Rework command not found detection
As far as I can tell the case that dates back to the first version of
this function cannot ever have worked, but perhaps back then the expect
framework was set up differently enough that it did. Regardless, these
days a missing binary shows up as the binary exiting with code 127, not
an exception being thrown, so switch to that.