Allow to pass multiple kernel configurations to --<prefix>/kernel-config.
This allows to specify multiple custom kernels to build.
Note that this differs from --<prefix>/extra-kernel-configs in that extra-kernel-configs
always appends to any kernel that cheribuild chooses to build, while the kernel-config
option should override the entire build list. The first kernel given to --<prefix>/kernel-config
is treated as the default kernel, installed in /boot/kernel.