git: refactor parsing of branch information
Use `git status` to get local and remote branches. The --porcelain=v2
output is easier to parse, but since I had already written code to support
v1 output as well this means we don't require git 2.11 yet.
In the future it might make sense to raise the minimum git version to 2.14
to be able to use --autostash unconditionally.