Support ZFS images for CheriBSD
Add support for simple ZFS root filesystem images. The images contain a
single 5GiB pool (zroot) mounted at /.  They are enabled with the option
"--disk-image/rootfs-type zfs" ("ufs" is the default.)

In the process, alter the way we create /etc/fstab.  For entries that
depend on command line options, insert them directly and then amend
the template file (pycheribuild/files/cheribsd/fstab.in) for constant or
user-specific entries.

This has been tested on morello-purecap and amd64. The results should
work on any platform that uses the FreeBSD loader as kernels don't
typically contain the ZFS module. This means it will not work on RISC-V
unless the kernel configuration is altered to build ZFS in.