Allow registering config options by defining class vars
This requires somewhat ugly metaclass hackery to reset the dictionary of
per-class config options before the class is defined (since otherwise
we would end up modifying the parent class dictionary and therefore
adding the option to subclasses that should not have it).
However, it does allow removing the setup_config_options override
for classes that only do basic option registration and might eventually
allow simplifying the initialization logic such that config options are
registered as part of __init__(). We might also be able to enable the
attribute-error warning in pytype once all classes have been migrated
to the new config option method.