Fix missing attribute when build MFS kernel
Now that we are no longer calling setup_config_options() for the
base class of the multi arch targets, this attribute is no longer
defined. Previously it was defined because we inherited the (incorrect)
value from BuildCHERIBSD, but now we need to explicitly set it. While
doing so fix a duplicate assignment.