gdb: Work around lack of native macOS/arm64 target support
We don't actually care about having support for macOS targets, except
that it means the build system automatically gives us unprefixed file
names. However, we can't do that on macOS/arm64, so use a macOS/amd64
triple as the closest thing and manually symlink to hide the ugliness
somewhat from the user.