pre-commit: Run everything but commit message hook for manual runs
The default includes manual but overriding it means it's not present
(rather than always being implicitly present). The commit message hook
relies on it really being a push to figure out what commits to look at,
but the rest make sense to be able to run manually.

Whilst here, fix inconsistent whitespace in the file.