boot_cheribsd: Check loader prompt was used if needed to pick kernel
When booting an image with a non-default kernel on a platform using
loader rather than direct booting (e.g. Morello), expect needs to
interrupt the loader and select the alternate kernel before booting, but
we weren't checking we managed to do this, only that if we saw the
loader countdown we were able to interrupt it. Thus add a check to fail
if the boot gets to single user (mounting root) without having seen the
loader prompt.