Update the regex for the boot loader.
Upstream FreeBSD changed the autoboot countdown string in September of
2019 (merged to CheriBSD in late December 2022).  The existing regex
didn't match the new string, so test boots requesting a non-default
kernel on Morello (e.g. the purecap kernel) have been booting the
default (hybrid) kernel instead.  This updates the regex to match on
the new string as well as the old.