cheribsd: Fix for .with_suffix replacing an existing suffix
In the kernel.FOO.full case, we'd end up with a dest_elf_with_dbg of
just kernel.full, so append rather than replace, and always copy the
kernel.full files when pretending, ignoring whether they exist, so that
this can easily be tested.

Also fix install_file to add force=True as used to be done for the
MFS-ROOT kernels (but not the replaced Jenkins-specific logic, since it
started from a clean build and thus never needed to bother).