Fix incorrect host dependency paths
When calling self.get_install_dir() we were previously ignoring the
cross_target argument since the suffixed target name would not be
resolved to a architecture-specific one. This should fix the build for
wayland where we were incorrectly searching for the wayland-scanner
binary in the sysroot instead of the host install dirs.