disk-image-minimal: Don't expect lib128(g) to have a full set of libraries
Both are minimal sets needed for SPEC CPU2006. Technically we should
still require libc, libc++ and related supporting libraries if the
directories are populated, but that's more invasive, and users aren't
expected to have lib128(g) at all, so if it's a bit liberal then it's
not really an issue.