SimpleProject: Teach maybe_strip_elf_file to check for symlinks
If a file is a symlink we don't want to actually strip it and lose the
link. Moreover, is_file() can raise a PermissionError for a symlink to a
path we can't access when building on the host (e.g. if you create an
extra-files tree with an absolute symlink to somewhere in /root)