libcxx: "Fix" shallow runtimes build for clones
For a shallow clone we have no way to know whether the commit in
question is present or not, so we have to just make some sensible
assumptions, i.e. that it's current CHERI-LLVM. This gives false
negatives, but we can't really do much better without looking in the
actual source for signs that the requirements are met.