Revert "libcxx: "Fix" shallow runtimes build for clones"
This is insufficient in Jenkins, where that commit happens to be present
but HEAD is shallow (at least on ctsrd-build-freebsd-f1), meaning we
don't hit the invalid commit case and instead end up thinking we know
it's not an ancestor. Moreover, in the case where you have a non-shallow
old checkout, that commit won't exist, yet we now treat it as True,
which is also wrong. Really contains_commit needs to get a lot more
complex to support handling these different cases with different
fallback values, but even then it won't work very well.

This reverts commit a8e1870f6e2546ee0ba6dd81ad2fa61a99c49a8b.