freebsd: Fix wrong cached objdir for --freebsd/toolchain=bootstrap
build_toolchain_root_dir ends up being called before our setup has
initialised the environment, and will read objdir when bootstrapping the
toolchain, so we need to ensure MAKEOBJDIRPREFIX is set otherwise it
will fall back on the build directory (which is an unhelpful default,
but that's another matter) and, more importantly, cache that incorrect
value for later uses, such as by the install step to enable the toor
user through the use of the pwd_mkdb tool, or by a disk-image target for
the various relevant bootstrap tools.

This is a bit of a hack. Ideally setup would add it to the environment
first, but it's overriding the environment rather than augmenting it,
and I don't know if that difference matters here. Ideally we'd switch to
adding to the environment and do it before setup, then this hack would
go away, but we'd need to be sure none of the parent classes are adding
unwanted things to the environment.