Add --implicit-mem-0 option to FreeRTOS demos to disable zero-init
Many RISC-V simulators start with memory implicitly zeroed. I have
PR-d a commit to https://github.com/CTSRD-CHERI/FreeRTOS-Demos-CHERI-RISC-V/pull/5
which allows the .bss and .sbss zero-initialization to be disabled
in these cases, because it takes too long to simulate and isn't necessary
in this case.

This commit adds an option to cheribuild which passes an option to
the FreeRTOS demos to disable the zero-initialization.