Bump Python minimum to 3.8 and rework CI a bit
Rather than testing every version, just maintain a Jenkins build for the
baseline and the latest. The latter can be done automatically as there's
a Docker image tagged latest, whilst the former needs a bit of manual
bumping. However, there should now be fewer places that need updating
with the version number.

Note that officially this means Ubuntu 18.04 is not supported. However,
for now, Jenkins is still using that for linux-baseline, and so we'll
need to keep it working in practice until we've migrated over to 20.04,
though probably we should jump straight to 22.04 for CI as 20.04 has a
bit under a year left to it as well.