Make sure we have a sensible HOME when running in Jenkins
Otherwise pip tries to install to /.local which goes as well as one
would expect.

Since this is messing with the tempsrc code, this also cleans up that. I
suspect what was previously happening was deleteDir() could not be
followed by something else inside the dir(...), as it wouldn't then
create the directory again.