tests: Shut pip up about breaking system packages
In their infinite wisdom the Python developers made even --user fail
despite being an extremely normal thing for people to use with no good
alternative.

Set PIP_BREAK_SYSTEM_PACKAGES=1 in the environment to shut it up once
and for all. --break-system-packages probably works too but I've already
wasted enough time dealing with this insanity and don't want to have to
figure out if that flag is too new for the baseline we support.