cheribsd: Add to rather than override environment after setup
We end up setting (X)CC etc in the environment before this point
specifically when using the bootstrap toolchain, since we use objdir in
build_toolchain_root_dir, and so by overriding the environment we lose
that, along with any others set by _setup_make_args. I can't obviously
see a reason why it has to be done this way, other than that it always
has been, dating back to when env_vars was handled in cheribsd.py before
it was abstracted out, and before other env vars were added to it. If
there are issues from this, they should be fixed elsewhere, not with
this dodgy approach.

Note that when not using the bootstrap toolchain we end up at this line
before calling _setup_make_args, so this is intended to align the
behaviour for the bootstrap toolchain with the non-bootstrap one.