cheribsd: Set MAKEOBJDIRPREFIX lazily pre-setup()
We can't just do it eagerly in __init__ since we need _init_called to be
set to true by Target.create_project for cheribsd-mfs-root-kernel (its
builddir will call cheribsd_reuse_build_dir, which then gets the base
cheribsd instance, which asserts that the caller is already
initialised).

This is all a bit of a mess still, as it seems that all the make_args
setting in setup() should be done in _setup_make_args(), with setup()
either calling _setup_make_args() if needed or just letting the laziness
do its job if not, but I don't want to untangle all that.