Add target for building bluespec-libraries, also 'bluespec' alias
'bluespec-libraries' builds and installs the B-lang-org/bsc-contrib repo
into the local install.

'bluespec' is a dummy target which depends on 'bluespec-libraries' which
depends on 'bluespec-compiler', so building 'bluespec' with dependencies
builds both.

Upstream bluespec-compiler's makefile tends to rebuild all its BSV files
even if nothing changed, so building 'bluespec-libraries' also causes the
BSV inside bluespec-compiler to get rebuilt.  Not sure if there's an easy
solution for this.