Add a --sysroot-install-dir-targets flag for jenkins-cheri-build
When target tarballs have been unpacked to the sysroot for use as
dependencies we need to tell jenkins-cheri-build not to assume they're
installed in the output root directory ($WORKSPACE/tarball). Sadly this
means piling more code into the existing hack.

Note that jenkins-cheri-build currently has the strange "feature" that
it always /opt/<suffix> regardless of what the target normally has, and
so we need to preserve this but relative to the sysroot, rather than
just not override the directory in the first place. This suffix also
includes the -for-<arch>-rootfs if present, unlike normal cheribuild.