Add a baseline test for jenkins hack rootfs behaviour
I am seeing CI errors when you enable hybrid-for-purecap-rootfs targets.
Add this test as a minimal reproducer showing the issue.
The problem is that the install() function in BuildKCoreAddons:
```
install_prefix = self.install_prefix
all_prefixes = " ".join(
    shlex.quote("/" + str(s.relative_to(self.rootfs_dir))) for s in self.dependency_install_prefixes
)
```
assumes that all dependencies share the same rootfs prefix, but this not
the case due to the implicit /opt/<target> directory being added even
for cheribsd targets if we don't specify an explicit install dir.