Consistently choose the same compiler
When we have multiple compilers with the same version such as
/usr/bin/clang and /usr/bin/clang-19 that point to the same underlying
file, the chosen compiler path depends on the directory iteration order.
Ensure we always pick the same one, preferring an unsuffixed compiler
and finally falling back to sorting by name.
This ensures that --reconfigure always picks up the same compiler.