disk-image-minimal: Use non-minimal METALOG for minimal's METALOG entries
Rather than synthesising our own approximation of what the real METALOG
lists for each entry we want to add, copy the real entry instead. This
means we don't have to meticulously set the ownership, permissions and
type (in particular, symlink or not) for every file, which we do in fact
get a bit wrong in some cases (especially for the PAM .so files, which
should be symlinks to the .so.6 files, not actual files.