Morello/baremetal: match Freestanding's sysroot dir
NewlibBaremetalTargetInfo targets should add the target triple to the
sysroot, as this was the layout expected by Newlib’s multilib install layout
(and GCC's sysroot). The current MorelloBaremetal targets shouldn't assume
a Newlib/GCC sysroot and should follow the FreeStanding target one. If there's
a new for GCC's sysroot and multilib layout, this should go to a new
NewlibMorelloBaremetalTargetInfo or something.