Linux/kernel: Add a target for building/running
This commit adds preliminary targets for cross compiling the Linux kerne
using LLVM, for AArch64 and RISC-V.
Unlike CheriBSD, Linux's userspace is scattered and depends on the project. For
instance, the same kernel built here could be used as a VM, for Yocto, Debian,
BusyBox, etc, so it's left to those projects to build, package, and run
userspace with this kernels.

* Morello: fetched from Arm's repos and only using hybrid kernel
* CHERI-Alliance: only builds a purecap kernel

A simple QEMULauch class is also added to package and run Linux:
* Morello: This builds and runs a minimal purecap busybox-based
  rootfs with muslc, then runs it on QEMU with networking enabled.
* CHERI-Alliance: Only builds and runs the kernel on QEMU until
  CHERI-RISC-V support for BusyBox and muslc are released.