qemu: Work around meson bug causing ld.lld to be probed as macOS ld64
If we decide to LTO, we use Clang+LLD, passing -flto in CFLAGS and -flto
plus -fuse-ld/--ld-path in LDFLAGS. However, the version of meson
vendored in QEMU that CHERI QEMU is based on only uses CFLAGS for linker
probing, and so ends up passing -flto without -fuse-ld/--ld-path. As a
result, it will try to do LTO with binutils, and thus needs the
corresponding libLLVMgold.so to be installed. For recent-ish versions of
LLVM packaged in Ubuntu, clang-N depends on llvm-N-linker-tools to
provide this, so it ends up believing GNU ld is the linker (which is
close enough to LLD), but on 20.04 with clang-10, as used in Morello CI,
there is no such dependency and so the attempt to probe the linker via
-Wl,--version gives an error of:

  /usr/bin/ld: /usr/lib/llvm-10/bin/../lib/LLVMgold.so: error loading plugin: /usr/lib/llvm-10/bin/../lib/LLVMgold.so: cannot open shared object file: No such file or directory
  clang: error: linker command failed with exit code 1 (use -v to see invocation)

The version of meson used then goes on to have the very strange
behaviour of treating that "(use -v to see invocation)" as a marker that
macOS ld64 is in use, and so happily reports that it probed macOS ld64.
This predictably falls over very quickly when corresponding linker flags
are added that LLD does not understand.

See: https://github.com/mesonbuild/meson/issues/6377