target_info: Fix building gdb-morello-* (and perhaps other targets)
When libiberty's configure script probes for headers like limits.h it
treats any warning as a no (despite not using -Werror), which includes
the unused argument warning for this linker flag. Ideally we would have
separated out compiler and linker flags for this interface, but that
would be a significant undertaking.

Fixes: d191c0db49a3 ("Enable -cheri-codeptr-relocs on Morello.")