jenkins_utils: Don't create projects for sysroot targets too early
Originally this code never created the project, instead creating a dummy
one temporarily to get the target_info, but this was changed to start
creating projects. However, instantiating_targets_should_warn is not yet
False, so that hits a RuntimeError for any target specified in
--sysroot-install-dir-targets.

Instead, wrap it in a ComputedDefaultValue (which is what its type
really should be anyway) so the concrete evaluation of the sysroot_dir
is deferred until we actually create the project later.

Fixes: e4e86b6096ad ("Remove cyclic import dependency to improve type checking")